Remove the use of SET ROWCOUNT and GOTO in incremental delete example
We aren't supposed to use SET ROWCOUNT in relation to DELETE statements any more (according to T-SQL documentation), and GOTO and labels are always a poor idea. It's much easier (and better practice) to just use TOP() with the DELETE to achieve the same outcome. So I've updated the example to do so. I also added statement terminators in the examples, and removed the use of regional-specific date formats that would only work in the USA. (Change to a date format that's not affected by region or language settings)