Update run-fiddler-trace.md
Could you please add the note "The log may contain personal information. You can review the log before you send them to us. You decide whether or not to send us the logs." on step 8? Some customers would like to recognize that the personal information can be collected during capturing fiddler and decide whether or not to collect it before running the .exe.