Introducing a new cause for this issue
Issue is not only affected by MFA conditional access policies. Any conditional access policy that requires user interaction will also cause the same behavior. 
Updating explanation for the issue and including screenshots.
Also, the previous resolution as MFA could be disabled for the end users, and the issue would still be present. The correct resolution is for the users to be excluded from CA policy that requires MFA.