Update directory-sync-tool-not-sync-object.md
Although already deprecated since 2017, we are now actually starting to block DirSync clients from synchronizing with Azure AD with the error referred on this page: "This version of the Microsoft Online Services Directory Sync tool is no longer supported."
The article was original meant to upgrade from 32-bit to 64-bit version of Directory Sync tool, but since the  64-bit version is also deprecated and will generate the same error message, I thought we could just re-purpose the same article for upgrades from DirSync tool to AADConnect. Hope this helps as we might start to see more support volume  about this error since we are now blocking all the deprecated DirSync tools.