Update cannot-access-mailbox.md
As mentioned in the article, the Onprem object must have the same name, alias, and user principal name (UPN) as the cloud mailbox. Updated the Syntax for the powershell example to reflect this. The original example fails as it sets the UPN as the Alias and causes the object to have incorrect properties.