Minor versions - can they be crawled or not?!
The current article says: 
"1. Verify that the version of the document that's missing from search results is a major version of the document. If the version is a minor version, it will not be displayed in the search results until it's checked in and published as a major version."

However here:
https://support.microsoft.com/en-us/office/draft-items-are-not-crawled-in-sharepoint-9198c307-13d6-425c-a174-542a60e410e4
it says that it does this by default but it could be changed: "By default SharePoint only crawls major versions of files and draft items are only viewable by their creators"

Can you please clarify?