Clarified opening text regarding device actions
We noticed due to the high volume of cases from support and requests from the customers regarding the device actions. In the documentation we are mentioning where we can find information in case administrator performed a remote action against a device but there is no information in case user initiated the request from company portal or portal.manage.microsoft.com. We stated that "most likely" if we cannot find the request under tenant admin > audit log is because user initiated. But we have a place in the console where the administrator can check if the user initiated the action through company portal or portal.manage.microsoft.com and that is under devices>monitor>device actions and I believe we should be documenting this so we can help both customers and support team avoid unnecessary troubleshooting.  Modified https://docs.microsoft.com/en-us/troubleshoot/mem/intune/troubleshoot-device-actions#how-do-i-tell-who-started-a-retirewipe line 44 and 63.