Update transfer-logins-passwords-between-instances.md
I had some issues with one of my customers due to the fact that the original sp_help_revlogin does not script the DEFAULT_LANGUAGE property of the Login. In LATAM is very common to have SQL Server installed in English but some of the Logins configured with Spanish as a Default Language. If we transfer this login without the DEFAULT_LANGUAGE property it will take English and this will create issues on the date format, etc. I propose a change in the sp_help_revlogin stored procedure to take the DEFAULT_LANGUAGE that the Login has and script it in the CREATE LOGIN so we can be sure that the Login is transfer as it is originally.