Adding more steps for FCI
Added some additional steps to prevent restarts on FCI when generating a dump. While working on a case, where customer had AG and FCI, the provided steps were not enough. The FCI healthcheck timeout was reached and Windows cluster restarted the instance, ignoring the property "If resource fails, do not restart"