Procedure 1 => i contains the wrong variable
under procedure step 1 => the wrong variable is used: %localappdata%. This lead to an error if you try to change into the folder: %localappdata%\AADConnect what does not exists. 
The correct path is usually: C:\ProgramData\AADConnect where the variable %ProgramData%  should be used. So the correct path is: %ProgramData%\AADConnect

I'd ask to commit this minor change.