Convert to version agnostic
Proposing changes to make the doc version-agnostic and the tags to show that it applies to all supported versions, removed one scenario that is only applicable for upgrading to SQL 2008 (original Scenario 2) and such upgrades are no longer supported, and fixed some formatting/grammatical issues. 

If it is not needed internally, I would also like to propose that we remove the "_Original product version:_ &nbsp; SQL Server 2008, SQL Server 2008 R2, SQL Server 2012 " line as it generates some confusion/doubts for customers.