A suggestion regarding 230s timeout
App Service 230 timeout does not happend on the Azure Load Balancer and the length of the timeout depends on the app's operating system. To prevent confusion, I would like to propose some modification like this request.
Please let me know if I am wrong in any of my recognitions. Thanks!