Add a reason for status is intermittently inactive
Based on support cases there is one more reason of why the SBC status is intermittently inactive.
If the installed root or intermediate certificate is not part of SBC certificate chain issuer. When the SBC start the three way handshake during authentication process the Teams service will not be able to validate the certificate chain on SBC and will reset the connection. The SBC might be able to authenticate again as soon as the public Root certificate is loaded again on service cache or the certificate chain is fixed on SBC.
•	Make sure that the intermediate and root certificate installed on SBC is correct.