Update troubleshoot-common-blue-screen-error.md
If rescue VM was created manually without the "az repair" extension, you can attempt to run the script manually by copying it from github and pasting it in the "run command" option of the rescue VM ONLY if the OS disk of VM in question is properly attached as data disk AND online. win-sacdump-on: https://raw.githubusercontent.com/Azure/repair-script-library/master/src/windows/sac-os-dump-enabler.ps1 

There is another option to capture dump which is not available on public documents. This helps when customer mannualy creates repair VM without any extension.