Adjusted the PowerShell script
PowerShell script incorrectly indicated that the MAC address would not change. It does change, so updated the wording. Also fixed one of the variables referenced toward the bottom of the script.

Lastly, for clarity, replicated the line from the portal steps that indicate that the VM will reboot to apply the change.