Update kernel-related-boot-issues.md
This entire document has been recreated. The previous document is not accurate (some of the kernel panics it was making a reference to, are not associated with no boot scenarios) and it was very outdated (recommending to delete the VM and recreated, instead of contemplating things like Repair VM, OS disk swap and so on and so forth. This document is needed by the Azure Linux Guided Troubleshooter.