Update linux-fulldisk-boot-error.md
I checked the 2nd suggestion, and the somewhat confusing flow of only pointing towards resolution 1 is appropriate.  Added some wording to clarify.  In this case, since   procedure 2 is already failing, the *only* thing possible is procedure 1, but once that is done procedure 2 should be retried.

The other areas were checked and modify or already look good as modifed.

#sign-off