Update server-software-support.md
We are adding support for upgrades on specific versions of Windows Server.  Preview of the new content is here: https://review.learn.microsoft.com/en-us/azure/virtual-machines/windows-in-place-upgrade?branch=pr-en-us-222834.  We should make a small note to this article to reflect the new support changes once the other article goes live.