Update in-place-system-upgrade.md
Support for upgrades from Windows Server 2016++ are coming to Azure soon.  Making a pull request adding in the supported scenario link, removing the Server OS list and remove the email alias which should no longer be used once the article goes live.  I've put in the current working link, it will need to be revised once the content is live.  The following OS versions are still not eligible, I'm removing them from this page and asking the tech writer for the new content to add them to the new page.

- Windows Server 2012 R2 Datacenter
- Windows Server 2012 R2 Standard
- Windows Server 2012 Datacenter
- Windows Server 2012 Standard
- Windows Server 2008 R2 Datacenter
- Windows Server 2008 R2 Standard