Update in-place-system-upgrade.md
Removing the Windows Server content from this article as well as the special alias.  We have a new Windows Server article for in-place upgrade.  Can we add a section for customers that points to it instead: https://learn.microsoft.com/en-us/azure/virtual-machines/windows-in-place-upgrade 

We have moved the unsupported Windows Server versions into this article and are killing off the alias.  I tried to submit this internally but it wouldnt let me so ping me with any questions (joscon)