Adding Note to call attention to the updated connector, for which this articles log details to not apply.
The new (unified) connector replaced the use of the two previous connectors in July of 2021, and apparently folks are trying to use this article to review details for that new connector. This should help divert them to the correct content.  This update was requested by the Intune dev/pm team.