Update common-errors-creating-and-assigning-flow-approvals.md
The GraphUserDetailByEmailMultipleFound error can also happen if the Requestor input resolves to more than one AAD object. 

Because the Requestor input is not displayed by default (you must click Show advanced options to see it), it can cause confusion since the error reported is not related to the identity of users in the Assigned To input.