Update turn-on-automatic-logon.md
I reordered when add the 'AutoAdminLogon' key to the beginning of the instructions, because in the 'DefaultPassword' section we make a NOTE about the 'AutoAdminLogon', but at that stage we hadn't even created or heard of it at that point in time. I am changing this because I just went through the instructions for my own server and got confused about what the 'AutoAdminLogon' even was as I was going through the directions step-by-step.