Create exports-fails-name-already-exits
We saw this behavior in one Cx instance and wanted to explain it for future cases. Note: This was not filed as a support ticket, but found through telemetry. 
Note: The chances of this happening should be very small, but rather than hoping it never happens (again), I decided to write this instead ;)