Fixing perfview instructions
Hey team. This article is delaying auto-attach support. Customers collect this trace but it only collects status monitor events but we need the other App Insights SDK events as well in order to really know if SDK attached, produced telemetry and sent it to ingestion correctly. Swapping out this instruction to pasted in the full provider list for AI SDK and auto-attach agent event providers.