Update enabling-smart-card-logon-third-party-certification-authorities.md
In example CDP/CRL URL, changed it from HTTPS to HTTP as HTTPS is not supported by any crypto API including Microsoft for CDP/CRL lookups. If you try to use HTTPS, you will end up in an endless loop of SSL trying to be verified thus will fail.

Reference:
RFC5280 https://datatracker.ietf.org/doc/html/rfc5280 CAs SHOULD NOT include URIs that specify https, ldaps, or similar schemes in extensions.  CAs that include an https URI in one of these extensions MUST ensure that the server's certificate can be validated without using the information that is pointed to by the URI.  Relying parties that choose to validate the server's certificate when obtaining information pointed to by an https URI in the cRLDistributionPoints, authorityInfoAccess, or subjectInfoAccess extensions MUST be prepared for the possibility that this will result in unbounded recursion.
Also other post/blogs:
https://learn.microsoft.com/en-us/answers/questions/169840/how-can-i-configure-cdp-with-https
https://www.sysadmins.lv/blog-en/designing-crl-distribution-points-and-authority-information-access-locations.aspx