Update troubleshoot-live-migration-issues.md
HPE is the largest Server OEM in multiple markets. HPE has customers hitting this issue such as with SR 2306140050000614. This is a known issue with bug 29849199 that has been fixed for Windows Server 2022 in bug 25909421. HPE asks that a support article be published to help convince customers to migrate to Windows Server 2022.