Adding TelemetryConfiguration leak Perf issue
We have many customers calling into App Insights and App Services who deal with performance issue when loading App Insights SDK. The behaviors are always tied to leaked TelemetryConfiguration objects. This is an effort to publicly document this behavior with guidance for customers to avoid these leaks and resolve performance issues.