Adding RG permissions FAQ
In the past month, we've seen two support cases on this where customer gave access to App Insights but users could not see webtest results. The reason is because they are separate child resources and have their own permissions. For optimal experience customers should give users read access to the resource group where App Insights is hosted and that will allow them to view the results of webtests.