Update redeploy-to-new-node-windows.md
I found the following in an "update" related page. It should be clearly mentioned in the doc to inform the readers:

"A virtual machine can encounter issues caused by user configurations or host infrastructure. If you have exhausted all other options under the troubleshooting menu, we suggest redeploying your virtual machine to a new host. "

Source: https://azure.microsoft.com/en-us/updates/use-azure-redeploy-functionality-to-transfer-virtual-machines-to-a-new-host/