Update redeploy-to-new-node-windows.md
I've added more information so readers know what will happen after redeployment, and that they know which disk is temporary and which one remains persistent.

"After you redeploy a VM, the original VM, including all the data that you saved on the temporary disk (such as drive D) and Ephemeral disk are deleted. However, the data stored on drive C and attached disks remains persisted. The configurations and associated resources are transferred to a new VM of the same size on a new host. The dynamic IP addresses associated with virtual network interface are updated."