Update error-556-insert-exec-failed.md
Removed text that this issue was fixed in SQL 2016 SP3, since this is contradicted by what is said later in the article. What actually is correct, I don't for sure, but in the thread https://learn.microsoft.com/en-us/answers/questions/1791795/sporadic-insert-exec-failed-because-the-stored-pro Naomi Nosonvosky seems to see this issue on SQL 2016 SP3.