Update capture-tcp-dump-linux-node-aks.md
in the section: https://learn.microsoft.com/en-us/troubleshoot/azure/azure-kubernetes/logs/capture-tcp-dump-linux-node-aks#step-5-transfer-the-capture-locally

the "kubectl cp" command assumes the file is located in the root path. adding a NOTE that mentions that the path may begin with "/host" if a "chroot /host" command was done when entering the debug pod.