Update in-place-system-upgrade.md
After testing the in-place upgrade in the lab, we confirmed that there will not be any changes to the source image when performing an in-place upgrade of the VM. While working on the case, the customer was looking for this information, and it is not updated in any public document. They were concerned about potential changes to the source image.