Update troubleshoot-connections-endpoints-same-virtual-network.md
The article looks good. I just changed a hyperlink which was pointing to this documentation: https://learn.microsoft.com/en-us/azure/aks/limit-egress-traffic?tabs=aks-with-system-assigned-identities#required-outbound-network-rules-and-fqdns-for-aks-clusters

This documentation was before larger and was splited in two parts. The more relevant part for our doc is https://learn.microsoft.com/en-us/azure/aks/outbound-rules-control-egress so I modified to redirect to this