Update vm-extension-provisioning-errors.md
When troubleshooting a case, I got into scenario where cx encountered OSProvisioningTimedOut for VMSS SCALE-in and cx referred our MS public doc to exactly check what would be timeout period and cx could not find any MS Public doc mentioning it and cx is very upset that our doc is not properly updated as this is very common error scenario .

So updated the public doc with below info with 2 scenarios of timeout periods as well which would be helpful for any cx's checking the same.



In some circumstances Instance VM scalein for VMSS, virtual machine deployment will fail with the message "OSProvisioningTimedOut". Most of the time, the virtual machine should be started and the operating system running. The Guest Agent should often report status as well.This error appears usually when the Provisioning Agent or the Guest Agent can't send status back to CRP but the virtual machine has been created and the operating system started. The timeout set for CRP to wait for the VM Guest OS to send a healthy signal is 40 minutes, if the VM Deployed includes Extensions the timeout is extended to wait for 90 minutes. This time might change at any time.