Update exclude-sites-from-retention-policy.md
Adding information about the 30-day grace period when sites are removed from retention. This is bypassed if sites are excluded, but not if they are removed. The grace period is referenced in another documentation article: https://learn.microsoft.com/en-us/purview/retention?tabs=table-overriden#releasing-a-policy-for-retention but it seems relevant to mention it here as well.