Remove redundant verification option 2 section
The old Msol commands under the "verification option 2" don't have an equivalent in Graph so it's no longer possible to retrieve AlternativeSecurityIds not LiveID (NetId).
I've asked 2 EEE teammates in the auth v-Team to review the Msol commands and their opinion is that "verification option 1" is sufficient for troubleshooting purposes, so I removed the option 2.