Update suse-public-cloud-connectivity-registration-issues.md
There are three more dependency packages which also need to be installed. While working on this collab task 2505050030003666001 with customer, I understand that when you run this command after copying the packages to affected VM it will fail to install them because of these missing packages(runc, containerd and libcontainers-sles-mounts)

sudo zypper --no-refresh --no-remote --non-interactive install --force *.rpm

Only after downloading them as well and moving them to that location of the affected VM, the installation went fine and the subsequently, VM able to establish the registration.